+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- `Enumerable` now throws `\LogicException` instead of `\DomainException`
+  when initialization failure.
+
 ## [0.2.0] - 2016-03-02
 ### Changed
 * (BC Break) `Enumerable` cannot be serialized anymore.